add reload apiserver
